requirements.txt

streamlit
pandas
altair
requests
python-dotenv
yfinance
openai
simpletransformers
torch
